requirements.txt
faster-whisper
sounddevice
numpy
openai
python-dotenv
scipy
watchdog
requests
pytest
